Added post storage
The upload form is now functional in the sense that posts are now saved to the posts table. This required adding post_type to post. This caused some bugs, solved by downgrading and remigrating the database with the new column. Images still require a table and to hold the reference to posts when stored.